+++ CHANGELOG.md
@@ -31,7 +31,7 @@ All non-core API/package changes (i.e. changes that do not affect the package de
 - `<ThemeProvider>` - Added custom `ThemeProvider` that uses `react-labs` theme by default.
 - `Theme` - Subtitle and body variations added.
 - `<Chip>` - Added component.
-- `<RefetchData>` - Added component to display data-fetching statuses.
+- `<RefetchData>` - Added component to display data-fetching statuses and last updated fetch.
 
 ### Fixed
 - `<DefaultTheme>` - Fixed theme object export.